Extract TraceVLA module-level helpers to pi0_trace_vla_base (§3 batch 1)
make_attn_mask / posemb_sincos / fourier_encode_2d were copy-pasted (identical modulo
docstrings) into all four trace/target model files. Move them into a new shared module
pi0_trace_vla_base.py (the eventual home of TraceVLABase); the four model files import them.

Pure functions, no parameters -> no flax param-tree change. Verified: the param-path fingerprint
is identical for all 8 trace/target configs vs. the pre-refactor baseline, and the trace_vla_lora
smoke matches (loss 2.4416). Committed atomically (base module + all four importers together) so
the tree never imports a missing symbol.
